+++ src/Request.php
@@ -62,18 +62,14 @@ class Request
         $created = time();
         $headers['(created)'] = strval($created);
         $headers['Date'] = date('D, d M Y H:i:s', $created-$offset) . ' GMT';
-        
+
         $headers['Host'] = $urlparts['host'];
         if (($urlparts['scheme'] == 'http'
-             // @codingStandardsIgnoreLine
-             && array_key_exists('port', $urlparts)
-             // @codingStandardsIgnoreLine
-             && $urlparts['port'] != 80)
+            && array_key_exists('port', $urlparts)
+            && $urlparts['port'] != 80)  
             || ($urlparts['scheme'] == 'https'
-                // @codingStandardsIgnoreLine
-                && array_key_exists('port', $urlparts)
-                // @codingStandardsIgnoreLine
-                && $urlparts['port'] != 443)
+            && array_key_exists('port', $urlparts)
+            && $urlparts['port'] != 443)
         ) {
             $headers['Host'] .= ':' . strval($urlparts['port']);
         }
@@ -119,37 +115,91 @@ class Request
     }
 
     /**
-     * Do an http/s request
+     * Do an http/s request async
      *
      * @param string $method  One of 'GET', 'POST', etc.
      * @param string $url     The URL to which to make the request
      * @param string $content Content to be sent to the server or null
      * @param string $ctype   The content type
      *
+     * @return A multicurl handle
+     */
+    private function _doAsync(
+        string $method,
+        string $url,
+        ?string $content,
+        ?string $ctype
+    ) {
+        $curl = $this->_do($method, $url, $content, $ctype, false);
+        $mh = curl_multi_init();
+
+        curl_setopt($curl, CURLOPT_SSL_VERIFYPEER, false);
+        curl_setopt($curl, CURLOPT_SSL_VERIFYHOST, false);
+
+        curl_multi_add_handle($mh, $curl);
+
+        do {
+
+            $mrc = curl_multi_exec($mh, $active);
+            $mrs = curl_multi_select($mh);
+
+            // not needed because of the usleep(15) below
+            // if ($mrs == -1) {
+            //     usleep(1);
+            // }
+
+            // normal curl takes 400+ ms
+            // socket writing takes ~ 100ms but is a lot of code overhead
+            // like https://stackoverflow.com/questions/14587514/php-fire-and-forget-post-request
+            // this takes ~ 180ms but is simple
+            // but weirdly, it only works when running through a proxy
+            // without a proxy, we need to wait for $active == 0 (the request to fully process)
+            // so we might as well then do a normal curl
+            // #TODO make PHP multithread?
+            usleep(15);
+
+
+        } while ($active == 1); //$mrs == 1 && ($mrc == CURLM_CALL_MULTI_PERFORM || $active == 1));
+
+        return $curl;
+    }
+
+    /**
+     * Do an http/s request
+     *
+     * @param string $method  One of 'GET', 'POST', etc.
+     * @param string $url     The URL to which to make the request
+     * @param string $content Content to be sent to the server or null
+     * @param string $ctype   The content type
+     * @param bool   $execute Whether or not to execute the curl or return the handle
+     *
      * @return An associative array containing 'status', 'content_type', and
      *         'content' or false
      */
     private function _do(
-        string $method, string $url,
-        ?string $content, ?string $ctype
+        string $method,
+        string $url,
+        ?string $content,
+        ?string $ctype,
+        bool $execute = true
     ) {
         $ret = false;
         $headers = array();
         $response = '';
 
-        curl_reset($this->_curl);
-        curl_setopt(
-            $this->_curl, CURLOPT_USERAGENT, 'ubiq-php/' . \Ubiq\VERSION
-        );
+        $curl = curl_init();
 
-        curl_setopt($this->_curl, CURLOPT_URL, $url);
+        curl_reset($curl);
+        curl_setopt($curl, CURLOPT_USERAGENT, \Ubiq\LIBRARY . '/' . \Ubiq\VERSION);
+
+        curl_setopt($curl, CURLOPT_URL, $url);
 
         if ($content) {
             array_push($headers, 'Expect:');
-            curl_setopt($this->_curl, CURLOPT_UPLOAD, true);
-            curl_setopt($this->_curl, CURLOPT_INFILESIZE, strlen($content));
+            curl_setopt($curl, CURLOPT_UPLOAD, true);
+            curl_setopt($curl, CURLOPT_INFILESIZE, strlen($content));
             curl_setopt(
-                $this->_curl, CURLOPT_READFUNCTION,
+                $curl, CURLOPT_READFUNCTION,
                 /* this function will consume $content */
                 function ($curl, $rsrc, $length) use (&$content) {
                     $ret = substr($content, 0, $length);
@@ -159,9 +209,9 @@ class Request
             );
         }
 
-        curl_setopt($this->_curl, CURLOPT_CUSTOMREQUEST, $method);
+        curl_setopt($curl, CURLOPT_CUSTOMREQUEST, $method);
         curl_setopt(
-            $this->_curl, CURLOPT_WRITEFUNCTION,
+            $curl, CURLOPT_WRITEFUNCTION,
             function ($curl, $data) use (&$response) {
                 $response .= $data;
                 return strlen($data);
@@ -178,9 +228,22 @@ class Request
             }
         }
 
-        curl_setopt($this->_curl, CURLOPT_HTTPHEADER, $headers);
+        curl_setopt($curl, CURLOPT_HTTPHEADER, $headers);
+
+        if (!$execute) {
+            return $curl;
+        }
+
+        $this->_curl = $curl;
 
         curl_exec($this->_curl);
+        $i = curl_getinfo($this->_curl);
+        $e = curl_error($this->_curl);
+
+        // print_r($i);
+        // print_r($e);
+        // echo $url;
+
         if (!curl_error($this->_curl)) {
             $ret = array(
                 'status' => curl_getinfo(
@@ -225,6 +288,21 @@ class Request
         return $this->_do('POST', $url, $content, $ctype);
     }
 
+    /**
+     * Do an http/s POST request asyncronously
+     *
+     * @param string $url     The URL to which to make the request
+     * @param string $content Content to be sent to the server or null
+     * @param string $ctype   The content type
+     *
+     * @return An associative array containing 'status', 'content_type', and
+     *         'content' or false
+     */
+    public function postAsync(
+        string $url, string $content, string $ctype = 'text/plain'
+    ) {
+        return $this->_doAsync('POST', $url, $content, $ctype);
+    }
     /**
      * Do an http/s PATCH request
      *